chore: hash_dict refactoring
