+++ coophive/utils.py
@@ -1,10 +1,11 @@
 """This module defines various utility classes and functions for the CoopHive simulator."""
 
+import hashlib
+import json
 from dataclasses import dataclass
 from enum import Enum
 
 from coophive.data_attribute import DataAttribute
-from coophive.hash_dict import hash_dict
 
 
 class AgentType(Enum):
@@ -72,3 +73,18 @@ extra_necessary_match_data = {
     "price_per_instruction": 1,
     "verification_method": "random",
 }
+
+
+def hash_dict(dict) -> str:
+    """Generate a SHA-256 hash of a dictionary.
+
+    Args:
+        dict (dict): Dictionary to be hashed.
+
+    Returns:
+        str: Hexadecimal representation of the SHA-256 hash.
+    """
+    hash_function = hashlib.sha256()
+    encoded = json.dumps(dict, sort_keys=True).encode()
+    hash_function.update(encoded)
+    return hash_function.hexdigest()
