+++ docs/img/packages_coophive.dot
@@ -9,7 +9,6 @@ charset="utf-8"
 "coophive.data_attribute" [color="#77AADD", label=<coophive.data_attribute>, shape="box", style="filled"];
 "coophive.deal" [color="#77AADD", label=<coophive.deal>, shape="box", style="filled"];
 "coophive.event" [color="#77AADD", label=<coophive.event>, shape="box", style="filled"];
-"coophive.hash_dict" [color="#77AADD", label=<coophive.hash_dict>, shape="box", style="filled"];
 "coophive.job" [color="#77AADD", label=<coophive.job>, shape="box", style="filled"];
 "coophive.job_offer" [color="#77AADD", label=<coophive.job_offer>, shape="box", style="filled"];
 "coophive.log_json" [color="#77AADD", label=<coophive.log_json>, shape="box", style="filled"];
@@ -47,7 +46,7 @@ charset="utf-8"
 "coophive.client" -> "coophive.smart_contract" [arrowhead="open", arrowtail="none"];
 "coophive.client" -> "coophive.solver" [arrowhead="open", arrowtail="none"];
 "coophive.client" -> "coophive.utils" [arrowhead="open", arrowtail="none"];
-"coophive.data_attribute" -> "coophive.hash_dict" [arrowhead="open", arrowtail="none"];
+"coophive.data_attribute" -> "coophive.utils" [arrowhead="open", arrowtail="none"];
 "coophive.deal" -> "coophive.data_attribute" [arrowhead="open", arrowtail="none"];
 "coophive.job_offer" -> "coophive.data_attribute" [arrowhead="open", arrowtail="none"];
 "coophive.machine" -> "coophive.data_attribute" [arrowhead="open", arrowtail="none"];
@@ -80,5 +79,4 @@ charset="utf-8"
 "coophive.solver" -> "coophive.resource_offer" [arrowhead="open", arrowtail="none"];
 "coophive.solver" -> "coophive.utils" [arrowhead="open", arrowtail="none"];
 "coophive.utils" -> "coophive.data_attribute" [arrowhead="open", arrowtail="none"];
-"coophive.utils" -> "coophive.hash_dict" [arrowhead="open", arrowtail="none"];
 }
